refactor(M): finished renaming from simulation to session for non ui stuff
