+++ frontend/messages/de/HomePage.json
@@ -32,7 +32,7 @@
     "title": "Ihre Vorteile",
     "description": "Alles, was Sie für Ihre erfolgreiche Führungskräfteentwicklung benötigen.",
     "items": {
-      "simulations": {
+      "sessions": {
         "title": "Realistische Übungsszenarien",
         "description": "Trainieren Sie mit einer KI, die sich Ihrem Kommunikationsstil anpasst."
       },
