+++ frontend/messages/en/HomePage.json
@@ -32,7 +32,7 @@
     "title": "Powerful Features for Leadership Development",
     "description": "Everything you need to master difficult workplace conversations and become a more effective leader.",
     "items": {
-      "simulations": {
+      "sessions": {
         "title": "Realistic Simulations",
         "description": "Practice with AI that responds naturally to your communication style and approach."
       },
