+++ frontend/src/app/[locale]/(standalone)/session/[id]/loading.tsx
@@ -1,4 +1,4 @@
-export default function SimulationPage() {
+export default function SessionPage() {
   return (
     <div className="flex flex-col h-[92vh] justify-between">
       {/* Custom Header */}
