+++ frontend/src/i18n/request.ts
@@ -17,7 +17,7 @@ export default getRequestConfig(async ({ requestLocale }) => {
     Login: (await import(`../../messages/${locale}/Login.json`)).default,
     Onboarding: (await import(`../../messages/${locale}/Onboarding.json`)).default,
     Preparation: (await import(`../../messages/${locale}/Preparation.json`)).default,
-    Simulation: (await import(`../../messages/${locale}/Simulation.json`)).default,
+    Session: (await import(`../../messages/${locale}/Session.json`)).default,
     Settings: (await import(`../../messages/${locale}/Settings.json`)).default,
     PrivacyPolicy: (await import(`../../messages/${locale}/PrivacyPolicy.json`)).default,
     TermsOfService: (await import(`../../messages/${locale}/TermsOfService.json`)).default,
