--- frontend/src/interfaces/models/Simulation.ts
@@ -1,8 +1,0 @@
-export enum ConnectionStatus {
-  Connecting = 'connecting',
-  Connected = 'connected',
-  Disconnected = 'disconnected',
-  Failed = 'failed',
-  Closed = 'closed',
-  New = 'new',
-}
